move color, emum, regex properties
